refactor: merge config
