+++ spras/config/util_enum.py
@@ -4,6 +4,10 @@ from typing import Any
 
 # https://stackoverflow.com/a/76883868/7589775
 class CaseInsensitiveEnum(str, Enum):
+    """
+    We prefer this over Enum to make sure the config parsing
+    is more relaxed when it comes to string enum values.
+    """
     @classmethod
     def _missing_(cls, value: Any):
         if isinstance(value, str):
